Elwinator/param section (#27)
* cmd/elwinator: add param section and weighted choice

This adds a section for creating new params. There is also a toggle for
wether the param should be weighted or uniform. Default is uniform. I
also refactored the code so there wasn't so much duplication.

* cmd/elwinator: fix imports

forgot to update imports when i refactored the code.

* cmd/elwinator: add choices to params

added the new actions and reducer for choices. also added a choice
editor component. started thinking about how to style everything. going
to use a form for the namespace and follow bootstrap form style.

* cmd/elwinator: add note for monday

* cmd/elwinator: huge code dump

* cmd/elwinator: routing and state remove old

worked on creating a basic routing workflow. have most of the pieces done.
need to finish choices and the other actions for each piece.

* cmd/elwinator: update urls

urls were pretty long i shortened them. probably should define them in a
file later on so we can edit them in one place instead of in every file.

* cmd/elwinator: update the markup for bootstrap

* cmd/elwinator: encode and refactor urls

i extracted the urls from the files so we could edit them in a single
location. also encode the params in the url so they appear correctly.

* cmd/elwinator: add the new choice markup

* cmd/elwinator: param gets correct data for url.

added the data needed to create the url for param pages.

* cmd/elwinator: add markup for new label component

* cmd/elwinator: add in bootstrap-theme css

* cmd/elwinator: add a Namespace List to root page

root page was pretty useless before

* cmd/elwinator: fix url in App

missed this url when i was updating them.

* cmd/elwinator: fix the button markup on NewLabel

* cmd/elwinator: update the look of namespace

* cmd/elwinator: add NewChoice functionality

* cmd/elwinator: basic creation completed.