add namespace to experiment response
previously we were not capturing the namespace on experiment responses.
This makes it difficult to differentiate experiments that use the same
name. For the majority of tests teams will not need to be aware of the
namespace they are using and it can be generated randomly. This will
assure that experiments that reuse names won't break the randomization.