cmd/elwin: migrated to viper for config
The configuration for elwin was previously only available through
environment variables. This meant that you had to know the environment
variables in order to properly configure elwin. Now you can use a file
of environment variables via the viper library. viper will search for
the file first in the current working directory, then in /etc/elwin.