all: better representation for param choices
before params had a Value struct with a value interface. I renamed this
to choices since it's not a single item. The interface is now also not
exported. The choices interface is satisfied with the following
signature.

    func Choice(uint64) string

I also updated weighted choices so they are not so prone to errors. If
the lengths of the weights and choices used to be different there would
be an error. This still needs some error checking for 0 values. Also
should update the storage protobufs so they have a similar layout.