change segments to []byte
segments used to be an array of size 16. this is very limiting on the
number of segments that an experiment can be assigned to. segments can
now be sized dynamically. storage will be responsible for ensuring all
segments in the same namespace are the same size. storage is also
responsible for ensuring segments are unique to experiments and there is
no overlap within a namespace.